Add model selection in java support. Only allied switch at this point is supported
