+++ subset/switches/src/main/java/switchtest/Main.java
@@ -1,11 +1,14 @@
 package switchtest;
 
+import switchtest.allied.AlliedTelesisX230;
+
 public class Main {
 
   public static void main(String[] args) throws Exception {
 
-    if (args.length != 3) {
-      throw new IllegalArgumentException("Expected ipAddress && port && supportPOE as arguments");
+    if (args.length != 4) {
+      throw new IllegalArgumentException(
+          "Expected ipAddress && port && supportPOE && switchModel as arguments");
     }
 
     String ipAddress = args[0];
@@ -14,8 +17,26 @@ public class Main {
 
     boolean supportsPOE = args[2].equals("true");
 
-    SwitchInterrogator switchInterrogator = new SwitchInterrogator(ipAddress, interfacePort, supportsPOE);
+    SupportedSwitchModelsEnum switchModel = null;
+    try {
+      SupportedSwitchModelsEnum.valueOf(args[3]);
+    } catch (Exception e) {
+      System.out.println(
+          "Unknown Switch Model: "
+              + args[3]
+              + "\nDefaulting to "
+              + SupportedSwitchModelsEnum.ALLIED_TELESIS_X230);
+      switchModel = SupportedSwitchModelsEnum.ALLIED_TELESIS_X230;
+    }
 
+    SwitchInterrogator switchInterrogator = null;
+    switch (switchModel) {
+      case CISCO_9300:
+        throw new Exception("Unsupported switch option");
+      case ALLIED_TELESIS_X230:
+      default:
+        switchInterrogator = new AlliedTelesisX230(ipAddress, interfacePort, supportsPOE);
+    }
     Thread switchInterrogatorThread = new Thread(switchInterrogator);
     switchInterrogatorThread.start();
   }
