+++ subset/switches/src/main/java/switchtest/SupportedSwitchModelsEnum.java
@@ -1,0 +1,7 @@
+package switchtest;
+
+public enum SupportedSwitchModelsEnum {
+
+    ALLIED_TELESIS_X230,
+    CISCO_9300
+}
